Fixed a bug in index.php with requirements status being determined by other requirements if the requirement did not have any testcases associated Removed config table in editMisc.php, since lite version was removed. Changed the link on index.php to showFullReq.php to point to showReqs.php instead.

+++ showReport.php
@@ -132,8 +132,13 @@ for ($i=0;$i<$num;$i++){
       
       //print_r ($output);
       $path = "RC/".$output[5]."/".$output[7]."/ss/".$output[9];
+      
         echo "<tr class='status_".$cl->status[$a]."'>";
+        if(file_exists($path)){
           echo "<td>--".$cl->action[$a]."</td><td><a target='_blank' href='$path'>".$dbh->getText('View Screenshot')."</a></td><td>".str_replace('&nbsp;',' ',$cl->var2[$a])."</td>";
+          }else{
+            echo "<td>--".$cl->action[$a]."</td><td>".str_replace('&nbsp;',' ',$cl->var2[$a])."</td><td></td>";
+          }
         echo "</tr>";
       }
   }
